Refactored files in a way such that the structure of the project aligns with CA
