+++ src/main/java/use_case/create_quiz/CohereAPI.java
@@ -1,4 +1,4 @@
-package use_case.CohereClient;
+package use_case.create_quiz;
 
 import com.cohere.api.Cohere;
 import com.cohere.api.requests.ChatRequest;
